refactor: Organize Angular components into subdirectories
Move each component into its own subdirectory for better organization:
- admin/ -> admin/dashboard/, admin/products/, admin/users/
- articles/ -> articles/article-detail/, articles/article-list/
- auth/ -> auth/login/, auth/register/
- forum/ -> forum/category-list/, forum/subcategory/, forum/thread-list/, forum/thread/
- products/ -> products/product-detail/, products/product-list/

Single-component folders (cart, contact, home, space) remain unchanged.
Update app.routes.ts lazy-load import paths accordingly.
