+++ frontend/src/app/features/admin/dashboard/admin-dashboard.scss
