+++ frontend/src/app/features/admin/products/admin-products.ts
