+++ frontend/src/app/features/articles/article-detail/article-detail.scss
