+++ frontend/src/app/features/articles/article-list/article-list.scss
