+++ frontend/src/app/features/auth/register/register.scss
