+++ frontend/src/app/features/auth/register/register.spec.ts
