+++ frontend/src/app/features/forum/category-list/forum-category-list.scss
