+++ frontend/src/app/features/forum/subcategory/forum-subcategory.scss
