+++ frontend/src/app/features/forum/thread-list/forum-thread-list.scss
