+++ frontend/src/app/features/forum/thread/forum-thread.scss
