+++ frontend/src/app/features/products/product-detail/product-detail.spec.ts
