+++ frontend/src/app/features/products/product-list/product-list.ts
